requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
httpx
python-multipart
yookassa
